+++ src/Exceptions/RateLimitException.php
@@ -1,6 +1,5 @@
 <?php
 
-
 namespace Nuwave\Lighthouse\Exceptions;
 
 use GraphQL\Error\ClientAware;
@@ -23,7 +22,6 @@ class RateLimitException extends RuntimeException implements ClientAware
      * Returns true when exception message is safe to be displayed to a client.
      *
      * @api
-     * @return bool
      */
     public function isClientSafe(): bool
     {
@@ -36,7 +34,6 @@ class RateLimitException extends RuntimeException implements ClientAware
      * Value "graphql" is reserved for errors produced by query parsing or validation, do not use it.
      *
      * @api
-     * @return string
      */
     public function getCategory(): string
     {